+++ clapi-hs.cabal
@@ -112,7 +112,7 @@ test-suite clapi-hs-test
                      , TestSerialisationProtocol
                      , PathSpec
                      , ProtocolSpec
-                     , TestNamespaceTracker
+                     , NamespaceTrackerSpec
                      , RelaySpec
                      , ServerSpec
                      , TestTextSerialisation
